+++ src/script.js
@@ -1,7 +1,7 @@
 import './styles.scss';
 import './scripts/blockMods/imageAttr';
 import './scripts/registerPlugin';
-import './scripts/helpers/validationHooks';
+import './scripts/validation/validationHooks';
 import './scripts/blockChecks/buttonValidation';
 import './scripts/blockChecks/imageValidation';
 import './scripts/blockChecks/tableValidation';